\name{C_pvalueclassify}
\alias{C_pvalueclassify}
\title{Calculating P-values for Medip-seq and MRE-seq data.}
\usage{
    C_pvalueclassify(type1, type2, type3, type4, sm1chring1, sm1chring2,
    sm1chring3, sm1chring4, p, typelength, sm1chringlength,
    pvalue=rep(0,length(sm1chring1)))
}
\description{
    Calculating P-values.
}
\arguments{
    \item{type1}{no.}
    \item{type2}{no.}
    \item{type3}{no.}
    \item{type4}{no.}
    \item{sm1chring1}{no.}
    \item{sm1chring2}{no.}
    \item{sm1chring3}{no.}
    \item{sm1chring4}{no.}
    \item{p}{no.}
    \item{typelength}{no.}
    \item{sm1chringlength}{no.}
    \item{pvalue}{no.}
}

\value{
    \item{problity}{no.}
}
\keyword{internal}
