\name{EM2_H1ESB1_MeDIP_sigleCpG}
\docType{data}
\alias{EM2_H1ESB1_MeDIP_sigleCpG}
\title{A simulation dataset of MeDIP CpG sites.}
\description{
    This data set gives 2000 CpG sites which include the chromosome of the
region, the start and the stop position of the MeDIP CpG sites.

}
\usage{EM2_H1ESB1_MeDIP_sigleCpG}

\format{A data.frame containing 2000 MeDIP CpG sites.}
\source{Zhang, B., Zhou, Y., et al. (2013). Functional DNA methylation
differences between tissues, cell types, and across individuals discovered
using the M&M algorithm. Genome Research. 23: 1522-1540.
}
\references{
    Zhou Y. (2018). Methylation-level Inferences and Detection of Differential
Methylation with Medip-seq Data.
}
\keyword{datasets}
\keyword{ internal }
