\name{SIMD-package}
\alias{SIMD-package}

\docType{package}
\title{
    A method to infer the methylation expression level for each CpG sites.
}
\description{
    SIMD is a package to infer the methylation expression level for each CpG
    sites.The main idea of SIMD is that by using statistical inference to with
    Medip-seq data method to infer the methylation level.
}

\author{
    Zhou Yan
    Maintainer:Zhou Yan <zhouy1016@szu.edu.cn>
}
\references{
    Zhou Y. (2018). Methylation-level inferences and detection of differential
methylation with Medip-seq data.}
\keyword{ package }
