\name{three_mre_cpg}
\docType{data}
\alias{three_mre_cpg}
\title{A simulation dataset of MRE CpG sites.}
\description{
    This data set gives 2000 CpG sites which include the chromosome of the
region, the start and the stop position of the MRE CpG sites.

}
\usage{three_mre_cpg}

\format{A data.frame containing 2000 MRE CpG sites.}
\source{Zhang, B., Zhou, Y., et al. (2013). Functional DNA methylation
differences between tissues, cell types, and across individuals discovered
using the M&M algorithm. Genome Research. 23: 1522-1540.
}
\references{
    Zhou Y. (2018). Methylation-level Inferences and Detection of Differential
Methylation with Medip-seq Data.
}
\keyword{datasets}
\keyword{ internal }
